refactor(core): consolidate semantic vector planning (#1110)
