+++ src/basic_memory/repository/search_repository_base.py
@@ -14,6 +14,11 @@ from sqlalchemy import Executable, Result, text
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from basic_memory import db
+from basic_memory.indexing.embedding_index_planning import (
+    EmbeddingIndexPlanner,
+    EntityVectorShardPlan as _EntityVectorShardPlan,
+)
+from basic_memory.repository import semantic_vector_sync
 from basic_memory.repository.embedding_provider import (
     EmbeddingProvider,
     embedding_provider_identity,
@@ -31,10 +36,8 @@ from basic_memory.repository.semantic_errors import (
     SemanticDependenciesMissingError,
     SemanticSearchDisabledError,
 )
-from basic_memory.repository import semantic_vector_sync
 from basic_memory.repository.semantic_vector_sync import (
     EntitySyncRuntime as _EntitySyncRuntime,
-    EntityVectorShardPlan as _EntityVectorShardPlan,
     PendingEmbeddingJob as _PendingEmbeddingJob,
     PreparedEntityVectorSync as _PreparedEntityVectorSync,
     VectorChunkState,
@@ -465,8 +468,8 @@ class SearchRepositoryBase(ABC):
         pending_records: list[VectorChunkRecord],
     ) -> _EntityVectorShardPlan:
         """Select the bounded shard to process for one entity sync invocation."""
-        return semantic_vector_sync.plan_entity_vector_shard(
-            pending_records,
+        return EmbeddingIndexPlanner().plan_entity_shard(
+            [record["chunk_key"] for record in pending_records],
             shard_size=OVERSIZED_ENTITY_VECTOR_SHARD_SIZE,
         )
